+++ albam/engines/mtframework/__init__.py
@@ -1,7 +1,7 @@
-from albam.mtframework.arc import Arc
-from albam.mtframework.mod import Mod156
-from albam.mtframework.tex import Tex112
-from albam.mtframework.mappers import FILE_ID_TO_EXTENSION, EXTENSION_TO_FILE_ID
+from albam.engines.mtframework.arc import Arc
+from albam.engines.mtframework.mod import Mod156
+from albam.engines.mtframework.tex import Tex112
+from albam.engines.mtframework.mappers import FILE_ID_TO_EXTENSION, EXTENSION_TO_FILE_ID
 
 
 __all__ = ('Arc', 'Mod156', 'Tex112', 'FILE_ID_TO_EXTENSION', 'EXTENSION_TO_FILE_ID')
