+++ albam/engines/mtframework/arc.py
@@ -3,7 +3,7 @@ import ntpath
 import os
 import zlib
 
-from albam.mtframework.mappers import FILE_ID_TO_EXTENSION, EXTENSION_TO_FILE_ID
+from albam.engines.mtframework.mappers import FILE_ID_TO_EXTENSION, EXTENSION_TO_FILE_ID
 from albam.utils import BaseStructure
