+++ tests/test_mtframework_blender.py
@@ -3,7 +3,7 @@ from itertools import chain
 
 import pytest
 
-from albam.mtframework.mod import Mod156, Mesh156
+from albam.engines.mtframework.mod import Mod156, Mesh156
 from albam.utils import get_offset
